Update and rename half.py to UserData.py
Here we are not retrieving the entire data what we get from the GitHub API we only select the attribute values what we required for analysis. That means by the process of retrieving itself we are cleaning the data. Here in this code, we get the users data from GitHub based on randomly generated User ID's